% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fPASS-package.R
\docType{package}
\name{fPASS-package}
\alias{fPASS}
\alias{fPASS-package}
\title{fPASS: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Salil Koner \email{salil.koner@duke.edu} (\href{https://orcid.org/0000-0003-1952-4210}{ORCID})

}
\keyword{internal}
