% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASS_ProjTest.R
\name{PASS_TS_ProjTest_sUfDA}
\alias{PASS_TS_ProjTest_sUfDA}
\title{Power and Sample size (PASS) calculation of
Two-Sample (TS) Projection-based test (ProjTest) for sparsely observed univariate functional data.}
\usage{
PASS_TS_ProjTest_sUfDA(
  target.SS,
  target.power,
  alpha.fix,
  alloc.ratio = c(1, 1),
  ngrid = 101,
  mean.diff,
  cov.type = c("ST", "NS"),
  cov.par,
  sigma2.e,
  nobs_per_subj,
  missing_type = c("nomiss", "constant"),
  missing_percent = 0,
  eigen.est.error,
  fpca_method = c("FPCA", "fpca.sc"),
  fpca_optns = list()
)
}
\arguments{
\item{target.SS}{Target sample size, must be a positive integer more than 3.}

\item{target.power}{Target power, must be a number between 0 and 1.
Only one of target.SS and target.power should be non-null. The
function will automatically solve for sample size if target.SS is NULL, and
solve for power if target.power is NULL.}

\item{alpha.fix}{Level of significance of the test. Must be a number between 0 and 1, possibly less than 0.2.}

\item{alloc.ratio}{Allocation of total sample size into the two groups. Must set as a vector of two
positive numbers. For equal allocation it should be put as c(1,1), for non-equal
allocation one can put c(2,1) or c(3,1) etc.}

\item{ngrid}{The number of equidistant grid points between (0,1) at which the eigen-components are estimated/calculated.
The function internally creates a regular grid of points of length = ngrid between (0,1),
i.e. \code{tgrid = seq(0,1,length.out = ngrid)} to conduct the test. It must set as a positive
integer between 51L and 501L.}

\item{mean.diff}{The difference in the mean function between the two groups. Must be supplied as a function class.}

\item{cov.type}{The type of the covariance structure of the data, must be either of "ST" (stationary) or
"NS" (non-stationary)}

\item{cov.par}{The covariance parameters that needs to be specified.
If \code{cov.type='ST'} then, cov.par
must be a named list of two elements, "var" and "cor", where "var" denotes
the common variance of the observations, must be a positive number; and "cor" denotes
The correlation structure between the observations.
It must be provided in the form specified in the
available in the documentation of R package \pkg{nlme}. Check the package documentation for more details.
The argument of this function is passed onto the \code{\link[nlme:corMatrix]{nlme::corMatrix()}} to extract the subject-specific
covariance matrix.
If \code{cov.type='NS'} then, cov.par
must be a named list of two elements, "cov.obj" and "eigen.comp".
Only one of the "cov.obj" or "eigen.comp" must be non-null. If the "cov.obj" is
specified, then it must be either a bivariate function or a ngrid \times ngrid matrix,
containing the covariance between two timepoints in tgrid. Alternatively,
if the true eigenfunctions (and number of eigenfunctions, say K) are known,
then the user can specify that by specifying eigen.comp.
In this case, the eigen.comp must be a named list with three elements,
"eig.obj", "eig.val", and "eig.compute, that specifies whether eigenfunctions are
needed to be estimated from the data.
The object "eig.obj" can either a function that returns a vector of length K for fixed timepoint
or it can be a matrix of dimension ngrid by K where eig.mat(i,k) = phi_k(t_i) for i=1,2,..ngrid and k=1,..K.
And the quantity "eig.val" must be a numeric vector of length K with positive elements.}

\item{sigma2.e}{Measurement error variance, should be set as zero if there is no measurement error.}

\item{nobs_per_subj}{The number of observations per subject. Each element of it greater than 3.}

\item{missing_type}{The type of missing in the number of observations of the subjects. Can be one of
\code{missing_type = "nomiss"} for no missing observations
or \code{missing_type = "constant"} for constant
missing. Only supports \code{missing_type = "constant"} now. Note that, for
varying nobs_per_subj, missing_type must be set as "nomiss" and missing_type == 0.}

\item{missing_percent}{The percentage of missing at each observation points for each subject.}

\item{eigen.est.error}{Specifies whether target.SS should be used to estimate the eigenfunctions. Must be logical.
If \code{eigen.est.error = FALSE}, then the eigenfunctions are estimated based on
a sample of size 5000 (a pre-specified large number). When \code{eigen.est.error = TRUE} ,
the value of target.SS is used as the sample size to estimate the eigenfunctions. This
means that when we want to calculate the asymptotic power based on a small value target.SS then,
setting \code{eigen.est.error = TRUE} will lead to invoke the estimation error of eigencomponents
in the resulting power. As a result, the power calculation may give different result at each run.
If one does not want to invoke eigenfunction estimation error in the power calculation, \code{eigen.est.error}
should be set to FALSE. For sample size calculation, \code{eigen.est.error} is automatically set to FALSE.}

\item{fpca_method}{The method by which the FPCA is computed. Must be one of c("FPCA", "fpca.sc").}

\item{fpca_optns}{Additional options to be passed onto the \code{\link[fdapace:FPCA]{fdapace::FPCA()}} function in order
to estimate the eigencomponents required for the sample size calculation.
It must be a named list with two elements, "FVEthreshold"  and "methodSelectK".}
}
\value{
If target.SS is null, it returns the optimal sample size required for specified target.power.
Otherwise returns the empirical power for the sample size specified by target.SS.
}
\description{
This function \code{PASS_TS_ProjTest_sUfDA()} computes the power and sample size required to conduct
the projection-based test of mean function between two groups of longitudinal data
or sparsely observed functional data under a random irregular design, under
common covariance structure between the groups.
}
\details{
The projection-based test assumes the two groups of the data share the identical
covariance structure. We allow the covariance structure of the data as both parametric
and non-parametric. If the covariance structure is non-parametric, then, the user must specify the
the covariance function in the form of a function.
The projection-based test represents the sparsely observed
functions parsimoniously in terms of Karhunen-Loeve (KL) expansion and use the
functional principal component analysis scores to test the difference in the mean
function between two groups, by taking advantage of multivariate Hotelling-\eqn{T^2} test.
See Wang (2021) for more details of the testing procedure. We use the \code{FPCA()}
function in the R package \code{fdapace} to conduct the functional principal component analysis.
}
\examples{
set.seed(12345)
target.SS <- NULL; target.power <- 0.8; alpha.fix <- 0.05; mean.diff <- function(t) {t};
cor.str <- nlme::corExp(1, form = ~ time | Subject);
sigma2 <- 1; sigma2.e <- 0.25; nobs_per_subj <- 6;
missing_type <- "constant"; missing_percent <- 0.01;
fpca_optns <- list("FVEthreshold" = 0.95)
\dontrun{
Sample_size <- PASS_TS_ProjTest_sUfDA(target.SS = target.SS,
target.power = target.power, alpha.fix = alpha.fix,
ngrid = 101, mean.diff= mean.diff,
cov.par = list("var" = sigma2, "cor" = cor.str), sigma2.e = sigma2.e,
nobs_per_subj = nobs_per_subj, missing_type = missing_type,
missing_percent = missing_percent, fpca_optns =fpca_optns)
}

target.SS    <- NULL; target.power <- 0.8; alpha.fix <- 0.05;
alloc.ratio  <- c(1,1)
mean.diff    <- function(t) {1 * (t^3)};
eig.fun      <- function(t) { sqrt(2)*c(sin(2*pi*t), cos(2*pi*t))}
eigen.comp   <- list("eig.val" = c(1, 0.5), "eig.obj" = eig.fun, "eig.compute" = FALSE)
cov.par      <- list("cov.obj" = NULL, "eigen.comp" = eigen.comp)
sigma2.e     <- 0.001; nobs_per_subj <- 4:7;
missing_type <- "nomiss"; missing_percent <- 0;
fpca_method  <- "fpca.sc"
fpca_optns   <- list("FVEthreshold" = 0.95, "methodSelectK" = "FVE")
ngrid        <- 101

\dontrun{
Sample_size  <- PASS_TS_ProjTest_sUfDA(target.SS = target.SS, target.power = target.power,
                                       alpha.fix = alpha.fix, alloc.ratio = alloc.ratio,
                                       ngrid = ngrid, mean.diff = mean.diff,
                                       cov.type = "NS", cov.par = cov.par, sigma2.e = sigma2.e,
                                       nobs_per_subj = nobs_per_subj, missing_type = missing_type,
                                       missing_percent = missing_percent,
                                       fpca_method = fpca_method, fpca_optns =fpca_optns)
}
}
\references{
Wang, Qiyao (2021)
\emph{Two-sample inference for sparse functional data,  Electronic Journal of Statistics,
Vol. 15, 1395-1423} \cr
\doi{https://doi.org/10.1214/21-EJS1802}.
}
\author{
Salil Koner \cr Maintainer: Salil Koner
\email{salil.koner@duke.edu}
}
