% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjTest.R
\name{TS_ProjTest_sUfDA}
\alias{TS_ProjTest_sUfDA}
\title{Twp-Sample Projection-based test for sparsely observed univariate functional data}
\usage{
TS_ProjTest_sUfDA(
  Samp.Size,
  alloc.ratio = c(1, 1),
  ngrid = 101,
  mean.diff,
  cov.type = c("ST", "NS"),
  cov.par,
  sigma2.e,
  nobs_per_subj,
  missing_type = c("nomiss", "constant"),
  missing_percent = 0,
  fpca_method = c("face", "fpca.sc"),
  fpca_optns = list()
)
}
\arguments{
\item{Samp.Size}{The sample size (combined of two groups)
at which the power to be calculated}

\item{mean.diff}{The difference in the mean function between the two groups. Must be supplied as a function class.}

\item{sigma2.e}{Measurement error variance, should be left as NULL if there is no measurement error.}

\item{nobs_per_subj}{The number of observations per subject. Must be a positive integer greater than 3.}

\item{missing_type}{The type pf missing in the number of observations of the subjects.
Only supports \code{missing_type = "constant"} now.}

\item{missing_percent}{The percentage of missing at each observation points for each subject.}

\item{fpca_optns}{Additional options to be passed onto the \code{\link[face:face.sparse]{face::face.sparse()}} function in order
to estimate the eigencomponents required for the test.}

\item{sigma2}{The common variance of the observations. Only implementing common variance across all measurement points.}

\item{cor.str}{The correlation structure between the observations. It must be provided in the form specified in the
available in the documentation of R package \pkg{nlme}. Check the package documentation for more details.
The argument of this function is passed onto the \code{\link[nlme:corMatrix]{nlme::corMatrix()}} to extract the subject-specific
covariance matrix.}
}
\value{
Returns the p-value of the projection-based hypothesis test.
}
\description{
The function \code{TS_ProjTest_sUfDA()} conducts
the Two-sample projection-based test of significant difference in mean function
for twp groups sparsely observed univariate functional data
or sparsely observed functional data under a random irregular design, under
the assumption of common covariance structure between the two groups.
}
\details{
The projection-based test assumes the two groups of the data share the identical
covariance structure. The projection-based test represents the sparsely observed
functions parsimoniously in terms of Karhunen-Loeve (KL) expansion and use the
functional principal component analysis scores to test the difference in the mean
function between two groups, by taking advantage of multivariate Hotelling-\eqn{T^2} test.
See Wang (2021) for more details of the testing procedure. We use the \code{face.sparse()}
function in the R package \code{face} to conduct the functional principal component analysis.
}
\examples{
set.seed(12345)
Samp.Size <- 480; delta <- 0.78;
mean.diff <- function(t) {delta*(t^3)};
alloc.ration <- c(1,1); ngrid <- 101;
cor.str <- nlme::corCompSymm(0.5, form = ~ time | Subject);
sigma2 <- 1.5; sigma2.e <- 0.25; nobs_per_subj <- 4:7;
cov.par <- list("var" = sigma2, "cor" = cor.str);
missing_type <- "nomiss"; missing_percent <- 0;
fpca_optns  <- list("FVEthreshold" = 0.95)
\dontrun{
est.power   <- TS_ProjTest_sUfDA(Samp.Size, alloc.ratio = c(1,1),
                                 ngrid = ngrid, mean.diff = mean.diff,
                                 cov.type = "ST", cov.par = cov.par, sigma2.e = sigma2.e,
                                 nobs_per_subj = nobs_per_subj, missing_type = missing_type,
                                 missing_percent = missing_percent,
                                 fpca_method = fpca_method, fpca_optns =fpca_optns)
}
#' cor.str     <- nlme::corAR1(0.5, form = ~ time | Subject);
cov.par     <- list("var" = sigma2, "cor" = cor.str);
\dontrun{
est.power   <- TS_ProjTest_sUfDA(Samp.Size = Samp.Size,
                                 alloc.ratio = alloc.ratio,
                                 ngrid = ngrid, mean.diff = mean.diff,
                                 cov.type = "ST", cov.par = cov.par, sigma2.e = sigma2.e,
                                 nobs_per_subj = nobs_per_subj, missing_type = missing_type,
                                 missing_percent = missing_percent,
                                 fpca_method = fpca_method, fpca_optns =fpca_optns)
}
}
\references{
Wang, Qiyao (2021)
\emph{Two-sample inference for sparse functional data,  Electronic Journal of Statistics,
Vol. 15, 1395-1423} \cr
\doi{https://doi.org/10.1214/21-EJS1802}.
}
\author{
Salil Koner \cr Maintainer: Salil Koner
\email{salil.koner@duke.edu}
}
