% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fPASS-package.R
\docType{package}
\name{fPASS-package}
\alias{fPASS}
\alias{fPASS-package}
\title{\code{fPASS} package}
\description{
See the README on \url{https://github.com/SalilKoner/fPASS/blob/main/README.md}
}
\details{
R package \pkg{fPASS} is designed to perform the power and sample size analysis
for functional under a dense and sparse (random) design and longitudinal data. The
function can handle data from wide variety of covariance structure, can be parametric,
or non-parametric. The user have a lot of flexibility into tweaking the arguments of the function
to assess the power function of the test under different sampling design and covariance process
of the response trajectory, and for any arbitrary mean difference function. Overall, the
functionality of the module is quite comprehensive and includes all the different cases
considered in the NCSS PASS software. We believe that this software can be an effective
clinical trial design tools when considering the projection-based test as the primary
decision making method.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SalilKoner/fPASS}
  \item Report bugs at \url{https://github.com/SalilKoner/fPASS/issues}
}


See the primary function \code{\link[=PASS_Proj_Test_ufDA]{PASS_Proj_Test_ufDA()}} to compute
the power and sample size of the test.
}
\author{
\strong{Maintainer}: Salil Koner \email{salil.koner@duke.edu} (\href{https://orcid.org/0000-0003-1952-4210}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Sheng Luo \email{sheng.luo@duke.edu} [contributor, funder]
}

}
\keyword{internal}
