% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fPASS-package.R
\docType{package}
\name{fPASS-package}
\alias{fPASS}
\alias{fPASS-package}
\title{fPASS: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Salil Koner \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
