% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fPASS-package.R
\docType{package}
\name{fPASS-package}
\alias{fPASS}
\alias{fPASS-package}
\title{\code{fPASS} package}
\description{
See the README on
or \href{https://github.com/SalilKoner/fPASS}{GitHub}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SalilKoner/fPASS}
  \item Report bugs at \url{https://github.com/SalilKoner/fPASS/issues}
}

}
\author{
\strong{Maintainer}: Salil Koner \email{salil.koner@duke.edu} (\href{https://orcid.org/0000-0003-1952-4210}{ORCID})

}
\keyword{internal}
