% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASS_fPCA_2Sample.R
\name{PASS_UfPCA}
\alias{PASS_UfPCA}
\title{Power and Sample size calculation of
Projection-based test for difference between mean function of
two groups.}
\usage{
PASS_UfPCA(
  target.SS,
  target.power,
  alpha.fix,
  mean.diff,
  sigma2,
  cor.str,
  sigma2.e,
  nobs_per_subj,
  missing_type = "constant",
  missing_percent = NULL
)
}
\arguments{
\item{target.SS}{Target sample size, must be a positive interger more than 3.}

\item{target.power}{Target power, must be a number between 0 and 1.
Only one of target.SS and target.power should be non-null. The
function will automatically solve for sample size if target.SS is NULL, and
solve for power if target.power is NULL.}

\item{alpha.fix}{Level of significance of the test. Must be a number between 0 and 1, possibly less than 0.2.}

\item{mean.diff}{The difference in the mean function between the two groups. Must be supplied as a function class.}

\item{sigma2}{The common variance of the observations. Only implementing common variance across all measurement points.}

\item{cor.str}{The correlation structure between the observations. It must be provided in the form specified in the
available in the documentation of nlme R package. Check the package documentation for more details.}

\item{sigma2.e}{Measurement error variance, should be left as NULL if there is no measurement error.}

\item{nobs_per_subj}{The number of observations per subject. Must be a positive integer greater than 3.}
}
\value{
If target.SS is null, it returns the optimal sample size required for specified target.power.
Otherwise returns the empirical power for the sample size specified by target.SS.
}
\description{
Computes the power and sample size required for different covariance structure of the data.
}
\examples{
set.seed(12345)
target.SS <- NULL; target.power <- 0.8; alpha.fix <- 0.05; mean.diff <- function(t) {t};
cor.str <- corExp(1, form = ~ time | Subject); sigma2 <- 1; sigma2.e <- 0.25; nobs <- 8;
Sample_size <- PASS_UfPCA(target.SS, target.power, alpha.fix, mean.diff, sigma2, cor.str, sigma2.e, nobs)

}
\author{
Salil Koner \cr Maintainer: Salil Koner
\email{salil.koner@duke.edu}
}
